update september break

+++ Assets/Scripts/Shop/SO/ShopBuyItemSO.cs
@@ -8,6 +8,6 @@ public abstract class ShopBuyItemSO : ScriptableObject
     [field: SerializeField] public LocalizedString ItemName { get; private set; }
     [field: SerializeField] public LocalizedString ItemDescription { get; private set; }
     [field: SerializeField] public Sprite ItemIcon { get; private set; }
-    [field: SerializeField] public float Price { get; private set; }
+    [field: SerializeField] public int Price { get; private set; }
     public abstract void OnBuyItem();
 }
